Version 2.0.0

+++ PiaSample/PiaSample/Sources/UI/ApplePayIntegration.swift
@@ -23,7 +23,7 @@ extension AppNavigation: PKPaymentAuthorizationViewControllerDelegate {
                 return
         }
                 
-        let itemCost = NSDecimalNumber(value: orderDetails.amount.inNotes - Float(orderDetails.shippingCost))
+        let itemCost = NSDecimalNumber(value: orderDetails.amount.inNotes - Double(orderDetails.shippingCost))
         
         let request = PiaSDK.makeApplePayPaymentRequest(
             for: supportedApplePayNetworks,
